Flag publication names that contain a comma
An actually interesting and useful warning to be issued while parsing PubsBio.py! A comma in a participant's name completely confuses the formatting of the list of participants of a session. The work-around (from Paul) is to put the participant in the [participant change name] of the .cfg file. However, it would be good to know what names require this attention, before proofreading the whole pocket program.